arithmetic: Move `Elem` to its own submodulue.
Introduce a few `leak...less_safe` methods that we hope to refactor
away.

```
git difftool HEAD^1:src/arithmetic/bigint.rs \
                    src/arithmetic/bigint/elem.rs
```
